Move handleGetRoomEvent from room-util to get-room-channel.js
