+++ electron/channels/channels.js
@@ -1,8 +1,9 @@
 const { ipcMain } = require('electron')
 const { channels } = require('../../src/shared/constants')
-const { handleGetRoomEvent } = require('../lib/room-util')
 const { handleEvent } = require('./handle-event')
+
 const { handleGetHomeEvent } = require('./get-home-channel')
+const { handleGetRoomEvent } = require('./get-room-channel')
 
 /* GET_ROOM */
 ipcMain.on(channels.GET_ROOM, handleEvent(handleGetRoomEvent))
