+++ electron/lib/room-util.js
@@ -26,15 +26,3 @@ export const saveRoom = async (path, diograph) => {
   console.log('Saving room', path)
   return writeFilePromise(`${path}/diograph.json`, data)
 }
-
-/**
- * Event handler for GET_ROOM channel
- * @function
- * @param event {Object} - Event from frontend via ipcMain
- * @param params {Object} - Uses only address key which is the folder path
- * @return {Promise} Resolves with object with diograph key with diograph as Object
- */
-export const handleGetRoomEvent = (event, { address }) =>
-  new Promise((resolve, reject) => {
-    resolve({ channelName: 'GET_ROOM', responseObject: getRoom(address) })
-  })
